% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_manipulation.R
\name{cache_timestamp}
\alias{cache_timestamp}
\title{Get the timestamp of the internal version of the metadata}
\usage{
cache_timestamp()
}
\value{
A string showing date, time, and timezone.
}
\description{
This is the date and time when the version of metadata inside this package
was last updated. This applies the the internal geocoding cache as well.
}
