% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata_plotting.R
\name{cat_heatmap}
\alias{cat_heatmap}
\title{Plot heatmap to show relationship between two categorical variables}
\usage{
cat_heatmap(pubs, row_var, col_var, ...)
}
\arguments{
\item{pubs}{A data frame with at least these columns:
\describe{
  \item{journal}{Name of the journal of the paper.}
  \item{year}{Year when the paper was published.}
}
There must be one row per publication or per method or species for each title
if faceting by those. If facetting, then a column whose name is the value in
`facet_by` must be present.}

\item{row_var}{Variable for rows of the heatmap. Tidyeval is supported.}

\item{col_var}{Variable for columns of the heatmap.}

\item{...}{Extra arguments to pass to `heatmap`}
}
\value{
A base R heatmap is plotted to the current device.
}
\description{
For instance, are papers for certain techniques more likely to be in certain
journals? Is there an association between species and journal? This is just
for visualization. Use `fisher.test` to see if it's significant. I still wonder
if I should rewrite this with ggplot2, which is more work than base R in this
case.
}
