% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_docs.R
\docType{data}
\name{europe_countries}
\alias{europe_countries}
\title{List of European countries}
\format{
A character vector
}
\source{
foobar
}
\usage{
europe_countries
}
\description{
List of European countries
}
\keyword{datasets}
