% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geocode.R
\name{geocode_address}
\alias{geocode_address}
\title{Geocode addresses and institutions with Google's API}
\usage{
geocode_address(df, address_col)
}
\arguments{
\item{df}{A data frame that has a column for addresses.}

\item{address_col}{Column for addresses. Tidyeval is used here so it doesn't
need to be quoted.}
}
\value{
A data frame with all columns of df, but with city, state, and country
added. Only rows with valid geocoding results are returned.
}
\description{
I'm using Google's API even though it's not free because it works better than
OpenStreetMap when it comes to addresses and institution names.
}
