% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geocode.R
\name{geocode_city}
\alias{geocode_city}
\title{Get longitude and latitude of cities}
\usage{
geocode_city(sheet, existing = NULL, geocode_method = c("OSM", "Google"))
}
\arguments{
\item{sheet}{A data frame with columns "city", "state/province", and "country".}

\item{existing}{A data frame with earlier output of this function, so only
cities not already in the collection are geocoded.}

\item{geocode_method}{Which geocoding service to use. Must be either "OSM"
(OpenStreetMap) or "Google". If "Google", then an API key must be provided in
environment variable `GGMAP_GOOGLE_API_KEY`.}
}
\value{
A data frame like the internal data `lcm_city_gc`, with columns
"city", "city2" (city name is concatenated to state and country in case
multiple cities around the world have the same name), "state/province",
"country", and "geometry" (sfc_POINT for the longitudes and latitudes).
}
\description{
Get longitude and latitude of cities
}
