% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geocode.R
\name{geocode_inst_city}
\alias{geocode_inst_city}
\title{Geocode institutions and cities}
\usage{
geocode_inst_city(sheet, cache = TRUE, cache_location = ".")
}
\arguments{
\item{sheet}{The tibble read from Google Sheets, which has columns country,
state/province, city, and institution.}

\item{cache}{Logial, whether to cache.}

\item{cache_location}{Where to save the cache}
}
\value{
A sf data frame for longitude and latitude of cities
}
\description{
Get longitude and latitude of institutions and cities. If cache is used and
there're some institutions or cities not already in the cache, those will be
added.
}
