% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_manipulation.R
\name{get_pubs_df}
\alias{get_pubs_df}
\title{Get data frame with one publication per row}
\usage{
get_pubs_df(sheet, other_cols = NULL)
}
\arguments{
\item{sheet}{The tibble for the spreadsheet.}

\item{other_cols}{Character vector of other columns to include.}
}
\value{
A data frame with one row per publication.
}
\description{
For current era papers, one dataset occupies one row. This function only keeps
information about publications so each row is a distinct publication.
}
