% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{hist_bool}
\alias{hist_bool}
\title{Plot histogram for each value of a logical variable}
\usage{
hist_bool(pubs, col_use, binwidth = 365)
}
\arguments{
\item{pubs}{A data frame with at least these columns:
\describe{
  \item{journal}{Name of the journal of the paper.}
  \item{year}{Year when the paper was published.}
}
There must be one row per publication or per method or species for each title
if faceting by those. If facetting, then a column whose name is the value in
`facet_by` must be present.}

\item{col_use}{Which logical variable to plot. Tidyeval is supported.}

\item{binwidth}{Width of bins for the histogram in days.}
}
\value{
A ggplot2 object
}
\description{
Plots 3 histograms showing the number of publications per year for TRUE, FALSE,
and NA, with the histogram overlaid on top of a translucent one for all
values. There's one facet per row so it's easy to compare how things change
with time.
}
