% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{hist_bool_line}
\alias{hist_bool_line}
\title{Plot outlines of histogram for a logical variable}
\usage{
hist_bool_line(
  pubs,
  col_use,
  facet_by = NULL,
  ncol = 3,
  binwidth = 365,
  preprints = FALSE
)
}
\arguments{
\item{pubs}{A data frame with at least these columns:
\describe{
  \item{journal}{Name of the journal of the paper.}
  \item{year}{Year when the paper was published.}
}
There must be one row per publication or per method or species for each title
if faceting by those. If facetting, then a column whose name is the value in
`facet_by` must be present.}

\item{col_use}{Which logical variable to plot. Tidyeval is supported.}

\item{facet_by}{Name of a column for facetting.}

\item{ncol}{Number of columns in facetted plot.}

\item{binwidth}{Width of bins for the histogram in days.}

\item{preprints}{Logical, whether preprints should be included.}
}
\description{
Kind of like `hist_bool`, but instead of plotting TRUE, FALSE, and NA in 3
separate facets, it plots them as an outline overlaid on a translucent
histogram for all values. This is useful when facetting with another categorical
variable, such as programming language.
}
