% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_docs.R
\docType{data}
\name{inst_words}
\alias{inst_words}
\title{List of institution words}
\format{
A character vector
}
\source{
Some words I often find in department names, and in addition, country
and city names.
}
\usage{
inst_words
}
\description{
List of institution words
}
\keyword{datasets}
