% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_docs.R
\docType{data}
\name{lang_img}
\alias{lang_img}
\title{Programming language logos}
\format{
A data frame with programming language and path to image of the logos.
}
\source{
online
}
\usage{
lang_img
}
\description{
Programming language logos
}
\keyword{datasets}
