% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_docs.R
\docType{data}
\name{lcm_abstracts}
\alias{lcm_abstracts}
\title{LCM abstracts from PubMed search}
\format{
Data frame
}
\source{
The search term was "((laser capture microdissection) OR (laser microdissection)) AND ((microarray) OR (transcriptome) OR (RNA-seq))"
}
\usage{
lcm_abstracts
}
\description{
Abstracts of LCM literautre from PubMed API in a data frame, last updated on
2020-09-07
}
\keyword{datasets}
