% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_docs.R
\docType{data}
\name{lcm_dfm2}
\alias{lcm_dfm2}
\title{Document term matrix of LCM abstracts}
\format{
A `quanteda` `dfm` object
}
\source{
See the `lcm_text_mining.Rmd` file in the `data-raw` in the [GitHub
repo of this package](https://www.github.com/pachterlab/museumst) for how
this matrix was generated, including what the phrases and stopwords are and
what counts as a rare term.
}
\usage{
lcm_dfm2
}
\description{
Each term is a unigram, i.e. a word, except for common phrases. Stop words and
rare words were removed.
}
\keyword{datasets}
