% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_docs.R
\docType{data}
\name{lcm_tokens}
\alias{lcm_tokens}
\title{quanteda tokens for LCM abstracts}
\format{
A quanteda tokens object
}
\source{
See the `lcm_text_mining.Rmd` file in the `data-raw` in the [GitHub
repo of this package](https://www.github.com/pachterlab/museumst) for how
tokenization was performed.
}
\usage{
lcm_tokens
}
\description{
Common phrases and unigrams, without stop words, and stemmed.
}
\keyword{datasets}
