% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{make_image_labs}
\alias{make_image_labs}
\title{Construct HTML image labels for time lines}
\usage{
make_image_labs(image_path, width, description, description_width)
}
\arguments{
\item{image_path}{Character vector, paths to the images of interest. Should be NA for
items that don't have an image.}

\item{width}{Width of the images on the time line in pixels to be shown on the
plot.}

\item{description}{The text for the items to be shown beneath the images.
Should be a character vector of the same length as `image_path`.}

\item{description_width}{Width of the description text in characters to wrap
the text in the labels.}
}
\value{
A character vector of HTML code to plot the images and descriptions
with `geom_richtext`.
}
\description{
Internal, called by `plot_timeline`.
}
