% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{make_labs}
\alias{make_labs}
\title{Construct plotmath labels for time lines}
\usage{
make_labs(description, description_width, ref_number)
}
\arguments{
\item{description}{The text for the items to be shown beneath the images.
Should be a character vector of the same length as `image_path`.}

\item{description_width}{Width of the description text in characters to wrap
the text in the labels.}

\item{ref_number}{A string of the number for reference. If NA, then the
reference will not be included in the plot.}
}
\value{
A character vector of plotmath, to enable superscripts, italics, and
text wrapping.
}
\description{
Internal, called by `plot_timeline`.
}
