% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text_mining.R
\name{my_searchK}
\alias{my_searchK}
\title{Find the appropriate number of topics for stm}
\usage{
my_searchK(documents, K, prevalence = ~1, seed1 = 19, ...)
}
\arguments{
\item{documents}{Quanteda dfm object.}

\item{K}{Numeric vector of numbers of topics to try.}

\item{prevalence}{Formula of topic prevalence.}

\item{seed1}{Random seed to use.}

\item{...}{Other arguments to pass to \code{stm::stm}.}
}
\value{
A tibble with the metrics output by searchK, plus a column for the
fitted model for each value of K tried.
}
\description{
This function is just like stm's searchK, but it uses tidyverse and allows a
formula for prevalence. Adapted from Julia Silge's blog.
}
