% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text_mining.R
\name{plot_ec}
\alias{plot_ec}
\title{Plot exclusivity vs. semantic coherence for different K's}
\usage{
plot_ec(k_result, Ks)
}
\arguments{
\item{k_result}{output of my_searchK}

\item{Ks}{Numeric vector of number of topics to plot.}
}
\value{
A ggplot object
}
\description{
Adapted from Julia Silge's blog.
}
