% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_facets_color}
\alias{plot_facets_color}
\title{Color facet strips by a variable}
\usage{
plot_facets_color(p, strip_color, palette)
}
\arguments{
\item{p}{A ggplot object for the original facetted plot.}

\item{strip_color}{Categorical column in the data for the original plot to
use to color the facet strips. Tidyeval is used here.}

\item{palette}{A character vector of colors. Can be named to assign each
color to a value in `strip_color`.}
}
\value{
Nothing, the plot is printed to device.
}
\description{
Color facet strips by a variable
}
