% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text_mining.R
\name{plot_k_result}
\alias{plot_k_result}
\title{Plot the metrics for each K tried}
\usage{
plot_k_result(k_result)
}
\arguments{
\item{k_result}{output of my_searchK}
}
\value{
A ggplot object.
}
\description{
Adapted from Julia Silge's blog.
}
