% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text_mining.R
\name{plot_topic_prop}
\alias{plot_topic_prop}
\title{Plot expected topic proportion in the corpus}
\usage{
plot_topic_prop(stm_res, n_top = 5, x_max = 0.15)
}
\arguments{
\item{stm_res}{A `STM` object.}

\item{n_top}{Number of top words to show per topic.}

\item{x_max}{Max in the x axis so all words will show in the plot.}
}
\value{
A ggplot object
}
\description{
This is equivalent to `plot(stm_res, type = "symmary")`, but it uses `ggplot2`
and looks nicer. Adapted from Julia Silge's blog.
}
