% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text_mining.R
\name{plot_topic_words}
\alias{plot_topic_words}
\title{Plot probability of each word in each topic}
\usage{
plot_topic_words(stm_res, n_top = 10, ncol = 5)
}
\arguments{
\item{stm_res}{A `STM` object.}

\item{n_top}{Number of top words to show per topic.}

\item{ncol}{Number of columns in facetted plot.}
}
\value{
A ggplot object
}
\description{
Adapted from Julia Silge's blog.
}
