% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata_plotting.R
\name{plot_wordcloud}
\alias{plot_wordcloud}
\title{Plot word cloud for a column}
\usage{
plot_wordcloud(
  sheet,
  col_use = title,
  species_use = "all",
  year_min = NA,
  year_max = NA,
  other_stop_words = NULL
)
}
\arguments{
\item{sheet}{The sheet read in from Google Sheets as a tibble.}

\item{col_use}{Which column to break down into words for word cloud. Tidyeval
is supported.}

\item{species_use}{Which species to filter by.}

\item{year_min}{Minimum year. The range of years will be >= year_min and
< year_max.}

\item{year_max}{Maximum year.}

\item{other_stop_words}{A character vector for other stop words besides the
stop_words data frame that comes with tidytext.}
}
\value{
A html widget.
}
\description{
Plots a word cloud for columns in the metadata that don't have a controlled
vocabulary. I guess maybe I should start using a controlled vocabulary.
}
