% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{pubs_per_capita}
\alias{pubs_per_capita}
\title{Plot per capita data as choropleth or bar plot}
\usage{
pubs_per_capita(
  pubs,
  zoom = c("world", "europe", "usa"),
  plot = c("choropleth", "bar"),
  label_states = TRUE
)
}
\arguments{
\item{pubs}{A data frame with one row per publication and columns country and
for the US, also a column "state/province".}

\item{zoom}{Whether to plot the world map or only Europe (centered on Western
Europe and some Eastern European countries are partially cropped off) or only
the US.}

\item{plot}{Whether to plot choropleth or bar plot.}

\item{label_states}{If plotting the US, whether to label the states.}
}
\value{
A ggplot2 object.
}
\description{
For the entire world, Europe (for European countries tend to be smaller) and
states within the US.
}
