% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{pubs_per_cat}
\alias{pubs_per_cat}
\title{Number of publications per category}
\usage{
pubs_per_cat(
  pubs,
  category,
  n_top = NULL,
  isotype = FALSE,
  img_df = NULL,
  img_unit = NULL
)
}
\arguments{
\item{pubs}{A data frame at least with a column for the category of interest.}

\item{category}{Column name to plot. Tidyeval is supported. If it's species
or language, then img_df does not have to be supplied for isotype plot since
the images are supplied internally.}

\item{n_top}{Number of top entries to plot. Especially useful for isotype.}

\item{isotype}{Logical, whether to make isotype plot, like one icon stands for
a certain number of publications.}

\item{img_df}{A data frame with one column with the same name as `category`
and another column called `image_paths` for path to the images. Relative path
is fine since it will be internally converted to absolute path. This argument
can be left as NULL if category is species or language, since in these two
cases, the `img_df` is provided internally.}

\item{img_unit}{Integer, how many publications for one icon.}
}
\value{
A ggplot2 object.
}
\description{
I think it looks better when the bars are horizontal to make the category
names easier to read, as the names can be quite long. This will plot a bar
chart for the number of publications per category, sorted according to the
number of publications.
}
