% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_manipulation.R
\name{read_major_events}
\alias{read_major_events}
\title{Read the major events sheet}
\usage{
read_major_events()
}
\value{
A tibble. In the returned tibble, the image column will be the paths
to the images of interest on your computer for plotting time lines.
}
\description{
There's a column for image file names. Absence of file name means that I can't
legally use the image for free. The ones present in the image column can be
reused in this package for free. PNAS's website says permission is not required
for reuse of figures for non-commercial purposes in the Requesting Permissions
section: https://www.pnas.org/page/about/rights-permissions The other images
used are under some form of Creative Commons license, and a corresponding link
to the license is in the license column.
}
