% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_manipulation.R
\name{read_metadata}
\alias{read_metadata}
\title{Read the metadata from Google Sheets}
\usage{
read_metadata(
  sheet_use = c("Prequel", "smFISH", "Array", "ISS", "Microdissection", "No imaging",
    "Analysis", "Prequel analysis")
)
}
\arguments{
\item{sheet_use}{Name of the sheet(s) to read.}
}
\value{
A tibble for the sheet of interest.
}
\description{
To do: Cache and add an argument to update cache. Unlike the geocodes, this
won't be stored in the data of this package since it's much faster to download
the sheet than to geocode.
}
