% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_docs.R
\docType{data}
\name{sheet_fill}
\alias{sheet_fill}
\title{Color palette for sheets}
\format{
A named vector of colors; the names are the sheets for data collection
}
\source{
Brewer
}
\usage{
sheet_fill
}
\description{
To give each sheet, i.e. each category of technologies, a fixed color for a
unified look among figures.
}
\keyword{datasets}
