% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_docs.R
\docType{data}
\name{species_cols}
\alias{species_cols}
\title{Color palette for species}
\format{
A named vector of colors; the names are the species, including "Other".
}
\source{
Brewer
}
\usage{
species_cols
}
\description{
To give each common species a fixed color for a unified look among figures.
}
\keyword{datasets}
