% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_docs.R
\docType{data}
\name{species_img}
\alias{species_img}
\title{Species images for isotype plot}
\format{
A data frame with species Latin names and path to images.
}
\source{
online
}
\usage{
species_img
}
\description{
These are the sources of images that are not stated to be under public domain
found online with filter "free to share and use".
No changes were made to the images unless indicated. The author and license,
if found, are listed here as well.
}
\details{
\describe{
  \item{drosophila.jpg}{http://gompel.org/images-2/drosophilidae}
  \item{zebrafish.jpg}{https://thumbs.dreamstime.com/m/zebrafish-zebra-barb-danio-rerio-freshwater-aquarium-fish-isolated-white-background-50201849.jpg}
  \item{ciona.jpg}{http://www.habitas.org.uk/marinelife/tunicata/cioints.jpg}
  \item{xenopus.jpg}{https://en.wikipedia.org/wiki/African_clawed_frog#/media/File:Xenopus_laevis_02.jpg
    by Brian Gratwicke. License: https://creativecommons.org/licenses/by/2.0/
  }
  \item{celegans.jpg}{https://en.wikipedia.org/wiki/Caenorhabditis_elegans#/media/File:Adult_Caenorhabditis_elegans.jpg
    by Kbradnam at English Wikipedia. License: https://creativecommons.org/licenses/by-sa/2.5/
    A smaller version of the original is used here.
  }
  \item{arabidopsis.jpg}{http://parts.igem.org/wiki/images/b/bd/Plants_Arabidopsis_thaliana_400px.jpg}
  \item{skull.jpg}{http://pngimg.com/download/42558 License: https://creativecommons.org/licenses/by-nc/4.0/
    The original was compressed and converted to jpg here.
  }
  \item{platynereis.jpg}{https://en.wikipedia.org/wiki/Epitoky#/media/File:PlatynereisDumeriliiFemaleEpitoke.tif
    By Martin Gühmann. A smaller jpg version of the original is used here.
    License: https://creativecommons.org/licenses/by-sa/4.0/
  }
  \item{yeast.jpg}{https://en.wikipedia.org/wiki/Shmoo#/media/File:Shmoos_s_cerevisiae.jpg
    By Pilarbini. This is a smaller version of the original.
    License: https://creativecommons.org/licenses/by-sa/4.0/deed.en
  }
}
}
\keyword{datasets}
