% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_docs.R
\docType{data}
\name{stm_res}
\alias{stm_res}
\title{stm model of LCM abstracts}
\format{
A `stm` object
}
\source{
See the `lcm_text_mining.Rmd` file in the `data-raw` in the [GitHub
repo of this package](https://www.github.com/pachterlab/museumst) for how
this model was fitted.
}
\usage{
stm_res
}
\description{
The R package `stm` was used for topic modeling of LCM abstracts, with date
published and city as covariates for topic prevalence. Thirty five topics
were used.
}
\keyword{datasets}
