% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{test_year_bool}
\alias{test_year_bool}
\title{Test whether something is associated with time}
\usage{
test_year_bool(pubs, col_use)
}
\arguments{
\item{pubs}{A data frame with at least these columns:
\describe{
  \item{journal}{Name of the journal of the paper.}
  \item{year}{Year when the paper was published.}
}
There must be one row per publication or per method or species for each title
if faceting by those. If facetting, then a column whose name is the value in
`facet_by` must be present.}

\item{col_use}{Which logical variable to plot. Tidyeval is supported.}
}
\value{
A glm object is returned invisibly. The summary is printed to screen
}
\description{
Fits a logistic regression model with glm to use year to predict proportion
of a logical variable is TRUE, and tests whether beta is 0.
}
