% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_docs.R
\docType{data}
\name{usa_w_pop}
\alias{usa_w_pop}
\title{American map with Alaska and Hawaii moved and with state population estimates}
\format{
sf
}
\source{
urbnmapr; state population data is from census.gov
}
\usage{
usa_w_pop
}
\description{
American map with Alaska and Hawaii moved and with state population estimates
}
\keyword{datasets}
