% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text_mining.R
\name{word_prop_scatter}
\alias{word_prop_scatter}
\title{Compare word frequencies between current era and prequel}
\usage{
word_prop_scatter(
  pubs,
  col_use = title,
  era = sheet,
  n_top = 20,
  other_stop_words = NULL
)
}
\arguments{
\item{pubs}{A data frame at least with a column for the category of interest.}

\item{col_use}{Which column to break down into words for word cloud. Tidyeval
is supported.}

\item{era}{Which column indicates whether the item is from prequel or current era.
Can be character, factor, or logical. Tidyeval is supported.}

\item{n_top}{Number of top entries to plot. Especially useful for isotype.}

\item{other_stop_words}{A character vector for other stop words besides the
stop_words data frame that comes with tidytext.}
}
\value{
A ggplot2 object.
}
\description{
Compare word frequencies between current era and prequel
}
