% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_docs.R
\docType{data}
\name{xylims}
\alias{xylims}
\title{Bounds of longitude and latitude for European maps}
\format{
A named numeric vector
}
\source{
I just tried until I got something that looks good.
}
\usage{
xylims
}
\description{
This is to make contributions of Europe easier to see. Since the most crowded
area is Western Europe, this box centers on Western Europe while some countries
in Eastern Europe are cropped out.
}
\keyword{datasets}
