% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_docs.R
\docType{data}
\name{lcm_city_gc}
\alias{lcm_city_gc}
\title{City geocodes of LCM literature}
\format{
sf data frame with longitude and latitude in the geometry column.
}
\source{
Address was pulled from the PubMed API, and geocoded with the Google
geocoding API with my own API key.
}
\usage{
lcm_city_gc
}
\description{
Longitude and latitude of institutions of first authors of LCM literature as
of publication, for plotting on maps.
}
\keyword{datasets}
