% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata_plotting.R
\name{plot_timeline}
\alias{plot_timeline}
\title{Plot time line with images and descriptions of each event}
\usage{
plot_timeline(
  events_df,
  ys,
  width = 100,
  description_width = 20,
  expand_x = c(0.1, 0.1),
  expand_y = c(0.05, 0.05)
)
}
\arguments{
\item{events_df}{A data frame with at least these columns:
\describe{
  \item{image}{Paths to the images to use in the label. NA for items without
  an image.}
  \item{date_published}{A vector of Date objects for dates when the event of
  interest was published. Note that the actual time when those events occurred
  is most likely earlier, sometimes much earlier than the publication date,
  and the events might have become quite influential before their publication.
  But the publication date is the easiest way to get an actual date.}
  \item{description}{Short descriptions of the events. The plot won't look
  good if the descriptions are too long.}
}}

\item{ys}{A numeric vector of the y coordinates of the items. Since I don't
know how to implement the ggrepel algorithm to make sure that the labels don't
overlap for geom_richtext, I have to manually set the y coordinates to make
sure that the labels don't overlap and look good.}

\item{width}{Width of the labels in pixels. How the labels will look depends
on the size of the plot. Would be cool if this can be specified in terms of
units within the plot.}

\item{description_width}{Width of the description text in characters to wrap
the text in the labels.}

\item{expand_x}{A numeric vector of length 2 of the proportion to expand the
x axis on the left and the right. This is a way to manually make sure that the
labels are not cropped off at the edge of the plot.}

\item{expand_y}{Same as expand_x, but for the y axis.}
}
\value{
A ggplot2 object for the plot.
}
\description{
The time line is plotted horizontally with ticks for years. The labels with
images and descriptions are above and/or below that line.
}
