% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata_plotting.R
\name{pubs_per_cat}
\alias{pubs_per_cat}
\title{Number of publications per category}
\usage{
pubs_per_cat(pubs, category, isotype = FALSE, img_unit = 5)
}
\arguments{
\item{pubs}{A data frame at least with a column for the category of interest.
If `isotype = TRUE`, then there should also be a column for images called "image",
already read by magick as image pointers. Each row is a unique combination of
the publication and the category of interest.}

\item{category}{Column name to plot. Tidyeval is supported.}

\item{isotype}{Logical, whether to make isotype plot, like one icon stands for
a certain number of publications.}

\item{img_unit}{Integer, how many publications for one icon.}
}
\value{
A ggplot2 object.
}
\description{
I think it looks better when the bars are horizontal to make the category
names easier to read, as the names can be quite long. This will plot a bar
chart for the number of publications per category, sorted according to the
number of publications.
}
