% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{pubs_per_year}
\alias{pubs_per_year}
\title{Number of publications per year}
\usage{
pubs_per_year(pubs, facet_by = NULL)
}
\arguments{
\item{pubs}{A data frame with at least these columns:
\describe{
  \item{journal}{Name of the journal of the paper.}
  \item{year}{Year when the paper was published.}
}
There must be one row per publication or per method or species for each title
if faceting by those. If facetting, then a column whose name is the value in
`facet_by` must be present.}

\item{facet_by}{Name of a column for facetting.}
}
\value{
A ggplot2 object.
}
\description{
Plot bar plot of the number of publications per year. Preprints are excluded
as the dates are incoherent with those for published papers. I find facetting
makes the plot easier to read than filling with different colors.
}
