% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_manipulation.R
\name{read_metadata}
\alias{read_metadata}
\title{Read the metadata from Google Sheets}
\usage{
read_metadata(
  sheet_use = c("Prequel", "smFISH", "NGS wo MD", "ISS", "Microdissection",
    "No priori", "Other", "Analysis", "Prequel analysis"),
  cache = TRUE,
  cache_location = "./sheets_cache",
  update = FALSE
)
}
\arguments{
\item{sheet_use}{Name of the sheet(s) to read.}

\item{cache}{Logial, whether to cache.}

\item{cache_location}{Where to save the cache}

\item{update}{Logical, whether to update cache. If not, then the cache is read,
and if the cache is absent, then a file contained within
this package is read. This file within the package is updated with each
version of this package. If TRUE, then the data is directly downloaded from
Google Sheets, which is guaranteed to be up to date.}
}
\value{
A tibble for the sheet of interest.
}
\description{
To do: Cache and add an argument to update cache. Unlike the geocodes, this
won't be stored in the data of this package since it's much faster to download
the sheet than to geocode.
}
