% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_docs.R
\docType{data}
\name{species_img}
\alias{species_img}
\title{Species images for isotype plot}
\format{
A data frame with species Latin names and path to images.
}
\source{
online
}
\usage{
species_img
}
\description{
Species images for isotype plot
}
\keyword{datasets}
